refactor(svelte2tsx): 外部 import 書き換えを helpers/rewrite_external_imports.rs へ
ワークスペース外へ出る相対 import の書き換えを
`helpers/rewriteExternalImports.ts` 相当のファイルへ移す。

純粋な移動でロジック変更はなく、出力はバイト単位で不変。
