+++ crates/rsvelte_core/src/svelte2tsx/helpers/mod.rs
@@ -1,0 +1,3 @@
+//! Standalone transforms applied to the generated output.
+
+pub mod rewrite_external_imports;
