+++ crates/rsvelte_core/src/svelte2tsx/mod.rs
@@ -1,3 +1,4 @@
+pub mod helpers;
 pub mod interfaces;
 #[allow(
     clippy::inherent_to_string_shadow_display,
